requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytz==2019.3
gunicorn==20.0.4
whitenoise==5.2.0
django-heroku==0.3.1
dj-database-url==0.5.0
django-cors-headers==3.2.1
